backend/requirements.txt
django
psycopg2-binary
gunicorn
djangorestframework
django-cors-headers
djangorestframework-simplejwt
